Address code maintainability issues from PR review #7151
🔄 CODE DEDUPLICATION:
1. Eliminate duplicate IV calculation functions
   - Created shared s3_sse_utils.go with unified calculateIVWithOffset function
   - Removed duplicate calculateSSES3IVWithOffset from s3_sse_s3.go
   - Removed duplicate calculateIVWithOffset from s3_sse_kms.go
   - Both SSE-KMS and SSE-S3 now use the same proven IV offset calculation
   - Ensures consistent cryptographic behavior across all SSE implementations

📋 SHARED HEADER LOGIC IMPROVEMENT:
2. Refactor shouldSkipEncryptionHeader for better clarity
   - Explicitly identify shared headers (AmzServerSideEncryption) used by multiple SSE types
   - Separate SSE-specific headers from shared headers for clearer reasoning
   - Added isSharedSSEHeader, isSSECOnlyHeader, isSSEKMSOnlyHeader, isSSES3OnlyHeader
   - Improved logic flow: shared headers are contextually assigned to appropriate SSE types
   - Enhanced code maintainability and reduced confusion about header ownership

🎯 BENEFITS:
- DRY principle: Single source of truth for IV offset calculation (40 lines → shared utility)
- Maintainability: Changes to IV calculation logic now only need updates in one place
- Clarity: Header filtering logic is now explicit about shared vs. specific headers
- Consistency: Same cryptographic operations across SSE-KMS and SSE-S3
- Future-proofing: Easier to add new SSE types or shared headers

✅ VERIFICATION:
- All code compiles successfully
- Basic S3 API tests pass
- No functional changes - purely structural improvements
- Same security guarantees maintained with better organization
