+++ weed/s3api/s3api_object_handlers_copy.go
@@ -2177,12 +2177,16 @@ func shouldSkipEncryptionHeader(headerKey string,
 	dstSSEC, dstSSEKMS, dstSSES3 bool) bool {
 
 	// Define header type groups for easier management
-	isSSECHeader := headerKey == s3_constants.AmzServerSideEncryptionCustomerAlgorithm ||
+	// First, identify shared headers that are used by multiple SSE types
+	isSharedSSEHeader := headerKey == s3_constants.AmzServerSideEncryption
+	
+	// SSE-C specific headers
+	isSSECOnlyHeader := headerKey == s3_constants.AmzServerSideEncryptionCustomerAlgorithm ||
 		headerKey == s3_constants.AmzServerSideEncryptionCustomerKeyMD5 ||
 		headerKey == s3_constants.SeaweedFSSSEIV
 
-	isSSEKMSHeader := headerKey == s3_constants.AmzServerSideEncryption ||
-		headerKey == s3_constants.AmzServerSideEncryptionAwsKmsKeyId ||
+	// SSE-KMS specific headers (excluding shared headers)
+	isSSEKMSOnlyHeader := headerKey == s3_constants.AmzServerSideEncryptionAwsKmsKeyId ||
 		headerKey == s3_constants.SeaweedFSSSEKMSKey ||
 		headerKey == s3_constants.SeaweedFSSSEKMSKeyID ||
 		headerKey == s3_constants.SeaweedFSSSEKMSEncryption ||
@@ -2190,10 +2194,16 @@ func shouldSkipEncryptionHeader(headerKey string,
 		headerKey == s3_constants.SeaweedFSSSEKMSEncryptionContext ||
 		headerKey == s3_constants.SeaweedFSSSEKMSBaseIV
 
-	isSSES3Header := headerKey == s3_constants.SeaweedFSSSES3Key ||
+	// SSE-S3 specific headers (excluding shared headers)  
+	isSSES3OnlyHeader := headerKey == s3_constants.SeaweedFSSSES3Key ||
 		headerKey == s3_constants.SeaweedFSSSES3Encryption ||
 		headerKey == s3_constants.SeaweedFSSSES3BaseIV ||
 		headerKey == s3_constants.SeaweedFSSSES3KeyData
+	
+	// Convenience groupings for backward compatibility and cleaner logic
+	isSSECHeader := isSSECOnlyHeader
+	isSSEKMSHeader := isSSEKMSOnlyHeader || (isSharedSSEHeader && (srcSSEKMS || dstSSEKMS))
+	isSSES3Header := isSSES3OnlyHeader || (isSharedSSEHeader && (srcSSES3 || dstSSES3))
 
 	// If it's not an encryption header, don't skip it
 	if !isSSECHeader && !isSSEKMSHeader && !isSSES3Header {
